requirements.txt
django==3.1.1
channels==2.4.0
djangorestframework==3.12.1
requests==2.24.0 
